Simplify the startup. Cs
简化startup.cs，做封装
