+++ Blog.Core/Startup.cs
@@ -1,21 +1,11 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Reflection;
-using AspNetCoreRateLimit;
+﻿using AspNetCoreRateLimit;
 using Autofac;
-using Autofac.Extras.DynamicProxy;
-using Blog.Core.AOP;
 using Blog.Core.Common;
 using Blog.Core.Common.LogHelper;
 using Blog.Core.Extensions;
 using Blog.Core.Filter;
 using Blog.Core.Hubs;
-using Blog.Core.IServices;
 using Blog.Core.Middlewares;
-using Blog.Core.Model;
-using log4net;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
@@ -23,10 +13,9 @@ using Microsoft.AspNetCore.Server.Kestrel.Core;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
-using static Blog.Core.SwaggerHelper.CustomApiVersion;
+using System.Reflection;
 
 namespace Blog.Core
 {
@@ -34,8 +23,6 @@ namespace Blog.Core
     {
 
         private IServiceCollection _services;
-        private List<Type> tsDIAutofac = new List<Type>();
-        private static readonly ILog log = LogManager.GetLogger(typeof(Startup));
 
         public Startup(IConfiguration configuration, IWebHostEnvironment env)
         {
@@ -107,101 +94,14 @@ namespace Blog.Core
             _services = services;
         }
 
-        // 注意在CreateDefaultBuilder中，添加Autofac服务工厂
+        // 注意在Program.CreateHostBuilder，添加Autofac服务工厂
         public void ConfigureContainer(ContainerBuilder builder)
         {
-            var basePath = AppContext.BaseDirectory;
-            //builder.RegisterType<AdvertisementServices>().As<IAdvertisementServices>();
-
-
-            #region 带有接口层的服务注入
-
-
-            var servicesDllFile = Path.Combine(basePath, "Blog.Core.Services.dll");
-            var repositoryDllFile = Path.Combine(basePath, "Blog.Core.Repository.dll");
-
-            if (!(File.Exists(servicesDllFile) && File.Exists(repositoryDllFile)))
-            {
-                var msg = "Repository.dll和service.dll 丢失，因为项目解耦了，所以需要先F6编译，再F5运行，请检查 bin 文件夹，并拷贝。";
-                log.Error(msg);
-                throw new Exception(msg);
-            }
-
-
-
-            // AOP 开关，如果想要打开指定的功能，只需要在 appsettigns.json 对应对应 true 就行。
-            var cacheType = new List<Type>();
-            if (Appsettings.app(new string[] { "AppSettings", "RedisCachingAOP", "Enabled" }).ObjToBool())
-            {
-                builder.RegisterType<BlogRedisCacheAOP>();
-                cacheType.Add(typeof(BlogRedisCacheAOP));
-            }
-            if (Appsettings.app(new string[] { "AppSettings", "MemoryCachingAOP", "Enabled" }).ObjToBool())
-            {
-                builder.RegisterType<BlogCacheAOP>();
-                cacheType.Add(typeof(BlogCacheAOP));
-            }
-            if (Appsettings.app(new string[] { "AppSettings", "TranAOP", "Enabled" }).ObjToBool())
-            {
-                builder.RegisterType<BlogTranAOP>();
-                cacheType.Add(typeof(BlogTranAOP));
-            }
-            if (Appsettings.app(new string[] { "AppSettings", "LogAOP", "Enabled" }).ObjToBool())
-            {
-                builder.RegisterType<BlogLogAOP>();
-                cacheType.Add(typeof(BlogLogAOP));
-            }
-
-            // 获取 Service.dll 程序集服务，并注册
-            var assemblysServices = Assembly.LoadFrom(servicesDllFile);
-            builder.RegisterAssemblyTypes(assemblysServices)
-                      .AsImplementedInterfaces()
-                      .InstancePerDependency()
-                      .EnableInterfaceInterceptors()//引用Autofac.Extras.DynamicProxy;
-                      .InterceptedBy(cacheType.ToArray());//允许将拦截器服务的列表分配给注册。
-
-            // 获取 Repository.dll 程序集服务，并注册
-            var assemblysRepository = Assembly.LoadFrom(repositoryDllFile);
-            builder.RegisterAssemblyTypes(assemblysRepository)
-                   .AsImplementedInterfaces()
-                   .InstancePerDependency();
-
-            #endregion
-
-            #region 没有接口层的服务层注入
-
-            //因为没有接口层，所以不能实现解耦，只能用 Load 方法。
-            //注意如果使用没有接口的服务，并想对其使用 AOP 拦截，就必须设置为虚方法
-            //var assemblysServicesNoInterfaces = Assembly.Load("Blog.Core.Services");
-            //builder.RegisterAssemblyTypes(assemblysServicesNoInterfaces);
-
-            #endregion
-
-            #region 没有接口的单独类，启用class代理拦截
-
-            //只能注入该类中的虚方法，且必须是public
-            builder.RegisterAssemblyTypes(Assembly.GetAssembly(typeof(Love)))
-                .EnableClassInterceptors()
-                .InterceptedBy(cacheType.ToArray());
-            #endregion
-
-            #region 单独注册一个含有接口的类，启用interface代理拦截
-
-            //不用虚方法
-            //builder.RegisterType<AopService>().As<IAopService>()
-            //   .AsImplementedInterfaces()
-            //   .EnableInterfaceInterceptors()
-            //   .InterceptedBy(typeof(BlogCacheAOP));
-            #endregion
-
-
-            // 这里和注入没关系，只是获取注册列表，请忽略
-            tsDIAutofac.AddRange(assemblysServices.GetTypes().ToList());
-            tsDIAutofac.AddRange(assemblysRepository.GetTypes().ToList());
+            builder.RegisterModule(new AutofacModuleRegister());
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
-        public void Configure(IApplicationBuilder app, IWebHostEnvironment env, IBlogArticleServices _blogArticleServices, ILoggerFactory loggerFactory)
+        public void Configure(IApplicationBuilder app, IWebHostEnvironment env)
         {
             // Ip限流,尽量放管道外层
             app.UseIpRateLimiting();
@@ -212,22 +112,12 @@ namespace Blog.Core
             // 记录ip请求
             app.UseIPLogMildd();
             // 查看注入的所有服务
-            app.UseAllServicesMildd(_services, tsDIAutofac);
+            app.UseAllServicesMildd(_services);
 
-            #region Environment
             if (env.IsDevelopment())
             {
                 // 在开发环境中，使用异常页面，这样可以暴露错误堆栈信息，所以不要放在生产环境。
                 app.UseDeveloperExceptionPage();
-
-                //app.Use(async (context, next) =>
-                //{
-                //    //这里会多次调用，这里测试一下就行，不要打开注释
-                //    //var blogs =await _blogArticleServices.GetBlogs();
-                //    var processName = System.Diagnostics.Process.GetCurrentProcess().ProcessName;
-                //    Console.WriteLine(processName);
-                //    await next();
-                //});
             }
             else
             {
@@ -235,42 +125,15 @@ namespace Blog.Core
                 // 在非开发环境中，使用HTTP严格安全传输(or HSTS) 对于保护web安全是非常重要的。
                 // 强制实施 HTTPS 在 ASP.NET Core，配合 app.UseHttpsRedirection
                 //app.UseHsts();
-
             }
-            #endregion
 
-            #region Swagger
-            app.UseSwagger();
-            app.UseSwaggerUI(c =>
-            {
-                //根据版本名称倒序 遍历展示
-                var ApiName = Appsettings.app(new string[] { "Startup", "ApiName" });
-                typeof(ApiVersions).GetEnumNames().OrderByDescending(e => e).ToList().ForEach(version =>
-                {
-                    c.SwaggerEndpoint($"/swagger/{version}/swagger.json", $"{ApiName} {version}");
-                });
-
-                // 将swagger首页，设置成我们自定义的页面，记得这个字符串的写法：解决方案名.index.html
-                c.IndexStream = () => GetType().GetTypeInfo().Assembly.GetManifestResourceStream("Blog.Core.index.html");
-
-                if (GetType().GetTypeInfo().Assembly.GetManifestResourceStream("Blog.Core.index.html") == null)
-                {
-                    var msg = "index.html的属性，必须设置为嵌入的资源";
-                    log.Error(msg);
-                    throw new Exception(msg);
-                }
-
-                // 路径配置，设置为空，表示直接在根域名（localhost:8001）访问该文件,注意localhost:8001/swagger是访问不到的，去launchSettings.json把launchUrl去掉，如果你想换一个路径，直接写名字即可，比如直接写c.RoutePrefix = "doc";
-                c.RoutePrefix = "";
-            });
-            #endregion
-
-
-
+            // 封装Swagger展示
+            app.UseSwaggerMildd(() => GetType().GetTypeInfo().Assembly.GetManifestResourceStream("Blog.Core.index.html"));
+           
             // ↓↓↓↓↓↓ 注意下边这些中间件的顺序，很重要 ↓↓↓↓↓↓
 
+            // CORS跨域
             app.UseCors("LimitRequests");
-
             // 跳转https
             //app.UseHttpsRedirection();
             // 使用静态文件
@@ -278,7 +141,7 @@ namespace Blog.Core
             // 使用cookie
             app.UseCookiePolicy();
             // 返回错误码
-            app.UseStatusCodePages();//把错误码返回前台，比如是404
+            app.UseStatusCodePages();
             // Routing
             app.UseRouting();
             // 这种自定义授权中间件，可以尝试，但不推荐
@@ -291,6 +154,7 @@ namespace Blog.Core
             // 开启异常中间件，要放到最后
             //app.UseExceptionHandlerMidd();
 
+            // 性能分析
             app.UseMiniProfiler();
 
             app.UseEndpoints(endpoints =>
